requirements.txt
django==5.0.7
djangorestframework==3.15.2
psycopg2-binary==2.9.9
gunicorn==22.0.0
python-decouple==3.8
dj-database-url==2.2.0